// fe/src/lib/api/generated/model/callSlipOperationalResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { CallSlipDestinationTypeValue } from './callSlipDestinationTypeValue';
import type { CallSlipFormRevisionSummary } from './callSlipFormRevisionSummary';
import type { CallSlipPersonSummary } from './callSlipPersonSummary';
import type { CallSlipReferralSummary } from './callSlipReferralSummary';

export interface CallSlipOperationalResponse {
  course_year_snapshot: string;
  created_at: string;
  destination_type: CallSlipDestinationTypeValue;
  form_revision: CallSlipFormRevisionSummary;
  id: string;
  interview_ended_at: string | null;
  issued_by: CallSlipPersonSummary;
  issued_by_name_snapshot: string;
  other_destination: string;
  recorded_by: CallSlipPersonSummary | null;
  referral: CallSlipReferralSummary | null;
  report_at: string;
  student: CallSlipPersonSummary;
  student_name_snapshot: string;
  updated_at: string;
}